Add CPU/GPU selection in the inference notebook
Added an option to choose between using a GPU or a CPU for doing inference.
By default, the code will use a GPU if available, otherwise a CPU.

A device parameter was added to BicubicUpscaledBaseline as well.
If no device is specified, the code automatically selects one using the method above.