Added cudatoolkit to requirements
Some users get the following error when running inference on CPU:

`AssertionError: Torch not compiled with CUDA enabled`
See: https://github.com/worldstrat/worldstrat/issues/3#issue-1315570026

The issue might be caused due to the `cudatoolkit` not being installed implicitly on some platforms. To ensure the package is installed, as it should be, it is now explicitly listed as a requirement.